Move datetime-related code into separate module
